policy: Add AutoEnable config option
This option is used to configure policy for auto enabling adapters when
found. This includes adapters found on start as well as adapters that
are plugged in later on.

Some distributions use legacy "hciconfig hci0 up" in boot-scripts to
automatically enable controllers on start. But that is causing SMP
kernel code to not being properly initialized.

This patch allows distributions to automatically enable adapters in
race free manner even if no agent is provided eg. in login screen.