core/device: Fix not calling accept callback
For GATT based services accept is used to indicate that there is a new
connection available so it needs to be called immediately once ATT
connects so the driver can register for notification, etc.

This also renames probe_gatt_profile to add_gatt_service which is what
in fact it was doing.