shared/queue: Remove reference count from queue_entry
Since the queue itself has support for reference and queue_foreach
actually make use of it to prevent crashes when queue_destroy is
called from the callback.

Note: The intention of reference count was to protect the entry itself
in queue_foreach, it may still be possible to have a crash if the next
entry if freed but that should be considered an implementation bug.