tools/mgmt-tester: Fix Remove Device tests for LE
The timing of Add/Remove device is different in recent kernels so that
if they are sent in rapid succession no HCI command will be sent at
all. Since the relevant thing is that LE scan is disabled after the
test add a timeout to verify this instead of waiting for a specific
HCI command.