core/advertising: Fix not parsing data properly
Dictionary entries should contain their value component as variant
container which requires dbus_message_iter_recurse to access.

Note: This will break the application that were using this API up to
now, but since this was experimental it should still be possible to fix
it without having to bump the interface revision.