tools/mgmt-tester: Test add local name in scan rsp data
Appropriate HCI command is expected in setup section because
as soon as new local name is registered in the kernel the
HCI_OP_LE_SET_SCAN_RSP_DATA is triggered. Then when we pass
Add_local_name_to_Scan_rsp_data flag in Add advertising
command the HCI command is not triggered because
Scan response data has not changed.