shared/att: Fix handling 'Insufficent Authentication' error respone
>From Core Specification 4.2 Vol.3 Part C (page 375):

When the link is not encrypted, the error code "Insufficient
Authentication" does not indicate that MITM protection is required.

When unauthenticated pairing has occurred and the link is currently
encrypted, the error code "Insufficient Authentication" indicates that
MITM protection is required.

If LE Secure Connections authenticated pairing is required but LE
legacy pairing has occurred and the link is currently encrypted, the
service request shall be rejected with the error code "Insufficient
Authentication".