core/gatt-database: Send offset if set to ReadValue
If there remote is using Read Blob command as part of a Read Long
procedure the offset needs to be set in order for the application to
detect from where it shall start.

Note that with this it is also possible to detect what the chunk sizes
needs to be as the first offset indicates the maximum payload.