Remove deprecated UHID_FEATURE API
I believe I have identified an issue with the HID-over-GATT (HoG) where,
through hidraw, the HIDIOCGFEATURE does not work correctly.

The symptom is that the ioctl call returns immediately with bogus data,
before the Read Request to the peripheral has been completed.

I believe the issue is caused by the hog-lib.c registering a handler for
both UHID_FEATURE and UHID_GET_REPORT events, which in the uhid header
file turn out to be the same enum.

This causes the get_report() to get called first, it issues the Read
Request and waits for it's completion. After this the get_feature() is
immediately called with the same uhid message, which sends the
UHID_FEATURE_ANSWER in to the kernel with stale data, which then gets
returned to the hidraw caller.

I have fixed this by removing the get_feature() as it is unnecessary
anyway. See attached patch.

I have tested with against both old and new uhid API (kernels 3.8 and
4.4).