shared/gatt-client: Simplify included discovery
Instead of doing include per service in a depth-first fashion do it
in breadth-first like its done for primary and secondary since it is
now possible to insert the include handle, not just append at the end,
it is not longer necessary to fetch the attribute in order to have
included handle.

As a bonus this can also speed up the discovery procedure since it will
result in less traffic over the air.