shared/tester: Add option to print to monitor
This adds option -m/--monitor that can be used together with
tester_monitor to send protocol data to be decoded by btmon.

In addition to that this also logs the tester output into btmon since
that has support to store its output on file this can be quite
convenient for reporting.