src/adapter.c : Event callback registration optimization
After read info complete,too many event registration functions are not conducive to subsequent expansion.
If we want to add a new event, according to the previous logic, we must add an event registration function, the code is too redundant.
Therefore, I added a global time callback registration in this submission.
If you want to add event registration later, you only need to add the new event and event callback function to the global table.